requirements.txt
Brotli==1.0.9
click==8.0.3
colorama==0.4.4
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.3
Flask-Compress==1.10.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.2
pandas==1.4.1
plotly==5.6.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
tenacity==8.0.1
Werkzeug==2.0.3
